+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
 * Added
   * PnP editor - component's number, same as on the preview page
+  * Output CSV file: optional filename postfix
 * Changed
   * Output CSV summary file: number of elements column is optional
 * Deprecated